test(okf): extract assertNoOutput helper
